Fixed norminette errors, had to create a new file to acomodate all the functions
